# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: LandIt API
  description: LandIt 智能求职助手 API 接口文档
  version: 1.0.0
  contact:
    name: Azir
servers:
  - url: http://localhost:8080
    description: 开发环境
...
paths:
...
  /api/resumes:
    get:
      tags:
        - resumes
      summary: 获取简历列表
      operationId: getResumes
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ResumeStatus'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ResumeList'
    post:
      tags:
        - resumes
      summary: 创建空白简历
      operationId: createResume
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResumeRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Resume'
  /api/resumes/upload:
    post:
      tags:
        - resumes
      summary: 上传简历文件
      operationId: uploadResume
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ResumeDetail'
  /api/resumes/{id}:
    get:
      tags:
        - resumes
      summary: 获取简历详情
      operationId: getResumeDetail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ResumeDetail'
    put:
      tags:
        - resumes
      summary: 更新简历
      operationId: updateResume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateResumeRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ResumeDetail'
    delete:
      tags:
        - resumes
      summary: 删除简历
      operationId: deleteResume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Void'
  /api/resumes/{id}/primary:
    put:
      tags:
        - resumes
      summary: 设置主简历
      operationId: setPrimaryResume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Void'
  /api/resumes/{id}/suggestions:
    get:
      tags:
        - resumes
      summary: 获取优化建议
      operationId: getResumeSuggestions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ResumeSuggestionList'
  /api/resumes/{id}/suggestions/{suggestionId}/apply:
    post:
      tags:
        - resumes
      summary: 应用优化建议
      operationId: applyResumeSuggestion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: suggestionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ResumeDetail'
  /api/resumes/{id}/optimize:
    post:
      tags:
        - resumes
      summary: AI优化简历
      operationId: optimizeResume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptimizeResumeRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_OptimizeResume'
  /api/resumes/{id}/export:
    get:
      tags:
        - resumes
      summary: 导出简历PDF
      operationId: exportResume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/pdf:
              schema:
                type: string
                format: binary

  # ==================== 面试模块 ====================
  /api/interviews/sessions:
    post:
      tags:
        - interviews
      summary: 开始面试会话
      operationId: startInterviewSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartSessionRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_StartSession'
  /api/interviews/sessions/{sessionId}/answers:
    post:
      tags:
        - interviews
      summary: 提交回答
      operationId: submitAnswer
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitAnswerRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_SubmitAnswer'
  /api/interviews/sessions/{sessionId}/hints:
    get:
      tags:
        - interviews
      summary: 请求提示
      operationId: getHint
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
        - name: questionId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Hint'
  /api/interviews/sessions/{sessionId}/finish:
    post:
      tags:
        - interviews
      summary: 结束面试
      operationId: finishInterview
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinishSessionRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_FinishSession'
  /api/interviews/history:
    get:
      tags:
        - interviews
      summary: 获取面试历史
      operationId: getInterviewHistory
      parameters:
        - name: type
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/InterviewType'
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_InterviewHistoryPage'
  /api/interviews/{id}:
    get:
      tags:
        - interviews
      summary: 获取面试详情
      operationId: getInterviewDetail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_InterviewDetail'
  /api/interviews/questions:
    get:
      tags:
        - interviews
      summary: 获取题库
      operationId: getInterviewQuestions
      parameters:
        - name: type
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/InterviewType'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_InterviewQuestions'

  # ==================== 复盘模块 ====================
  /api/reviews:
    get:
      tags:
        - reviews
      summary: 获取复盘列表
      operationId: getReviews
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ReviewListPage'
  /api/reviews/{id}:
    get:
      tags:
        - reviews
      summary: 获取复盘详情
      operationId: getReviewDetail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_InterviewReview'
  /api/reviews/{id}/export:
    get:
      tags:
        - reviews
      summary: 导出复盘报告
      operationId: exportReview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/pdf:
              schema:
                type: string
                format: binary

  # ==================== 统计模块 ====================
...
components:
  schemas:
...
    ResumeStatus:
      type: string
      enum:
        - optimized
        - draft
    SuggestionType:
      type: string
      enum:
        - critical
        - improvement
        - enhancement
    InterviewType:
      type: string
      enum:
        - technical
        - behavioral
    InterviewStatus:
      type: string
      enum:
        - completed
        - in_progress
    QuestionDifficulty:
      type: string
      enum:
        - easy
        - medium
        - hard
...
    Resume:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        targetPosition:
          type: string
        updatedAt:
          type: string
        status:
          $ref: '#/components/schemas/ResumeStatus'
        score:
          type: integer
        completeness:
          type: integer
        isPrimary:
          type: boolean
    ResumeSection:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        title:
          type: string
        content:
          type: object
        score:
          type: integer
        suggestions:
          type: array
          items:
            $ref: '#/components/schemas/ResumeSuggestionItem'
    ResumeSuggestionItem:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/SuggestionType'
        content:
          type: string
    ResumeDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        targetPosition:
          type: string
        sections:
          type: array
          items:
            $ref: '#/components/schemas/ResumeSection'
        overallScore:
          type: integer
        keywordMatch:
          type: integer
        formatScore:
          type: integer
        contentScore:
          type: integer
    ResumeSuggestion:
      type: object
      properties:
        id:
          type: string
        type:
          $ref: '#/components/schemas/SuggestionType'
        category:
          type: string
        title:
          type: string
        description:
          type: string
        impact:
          type: string
        position:
          type: string
    CreateResumeRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        targetPosition:
          type: string
    UpdateResumeRequest:
      type: object
      properties:
        name:
          type: string
        sections:
          type: array
          items:
            $ref: '#/components/schemas/ResumeSection'
    OptimizeResumeRequest:
      type: object
      properties:
        targetPosition:
          type: string
    OptimizeResumeResponse:
      type: object
      properties:
        optimizedResume:
          $ref: '#/components/schemas/ResumeDetail'
        suggestions:
          type: array
          items:
            $ref: '#/components/schemas/ResumeSuggestion'

    # ==================== 面试相关 ====================
    Interview:
      type: object
      properties:
        id:
          type: string
        type:
          $ref: '#/components/schemas/InterviewType'
        position:
          type: string
        company:
          type: string
        date:
          type: string
        duration:
          type: integer
        score:
          type: integer
        status:
          $ref: '#/components/schemas/InterviewStatus'
        questions:
          type: integer
        correctAnswers:
          type: integer
    InterviewQuestion:
      type: object
      properties:
        id:
          type: string
        category:
          type: string
        difficulty:
          $ref: '#/components/schemas/QuestionDifficulty'
        question:
          type: string
        followUp:
          type: string
        keyPoints:
          type: array
          items:
            type: string
        sampleAnswer:
          type: string
    InterviewQuestions:
      type: object
      properties:
        technical:
          type: array
          items:
            $ref: '#/components/schemas/InterviewQuestion'
        behavioral:
          type: array
          items:
            $ref: '#/components/schemas/InterviewQuestion'
    Conversation:
      type: object
      properties:
        id:
          type: string
        role:
          $ref: '#/components/schemas/ConversationRole'
        content:
          type: string
        timestamp:
          type: string
        score:
          type: integer
        feedback:
          type: string
    InterviewAnalysis:
      type: object
      properties:
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
        overallFeedback:
          type: string
    InterviewDetail:
      type: object
      properties:
        id:
          type: string
        type:
          $ref: '#/components/schemas/InterviewType'
        position:
          type: string
        company:
          type: string
        date:
          type: string
        duration:
          type: integer
        score:
          type: integer
        conversation:
          type: array
          items:
            $ref: '#/components/schemas/Conversation'
        analysis:
          $ref: '#/components/schemas/InterviewAnalysis'
    StartSessionRequest:
      type: object
      required:
        - type
        - position
      properties:
        type:
          $ref: '#/components/schemas/InterviewType'
        position:
          type: string
        difficulty:
          $ref: '#/components/schemas/QuestionDifficulty'
        questionCount:
          type: integer
    SessionQuestion:
      type: object
      properties:
        id:
          type: string
        question:
          type: string
        keyPoints:
          type: array
          items:
            type: string
    StartSessionResponse:
      type: object
      properties:
        sessionId:
          type: string
        firstQuestion:
          $ref: '#/components/schemas/SessionQuestion'
    SubmitAnswerRequest:
      type: object
      required:
        - questionId
        - answer
      properties:
        questionId:
          type: string
        answer:
          type: string
        timestamp:
          type: string
    SubmitAnswerResponse:
      type: object
      properties:
        score:
          type: integer
        feedback:
          type: string
        nextQuestion:
          $ref: '#/components/schemas/SessionQuestion'
        isFollowUp:
          type: boolean
        finished:
          type: boolean
    HintResponse:
      type: object
      properties:
        hint:
          type: string
        keyPoints:
          type: array
          items:
            type: string
    FinishSessionRequest:
      type: object
      properties:
        duration:
          type: integer
    FinishSessionResponse:
      type: object
      properties:
        interviewId:
          type: string
        reviewId:
          type: string
        overallScore:
          type: integer

    # ==================== 复盘相关 ====================
    ReviewDimension:
      type: object
      properties:
        name:
          type: string
        score:
          type: integer
        maxScore:
          type: integer
        feedback:
          type: string
...
    ImprovementPlan:
      type: object
      properties:
        category:
          type: string
        items:
          type: array
          items:
            type: string
    InterviewReview:
      type: object
      properties:
        id:
          type: string
        interviewId:
          type: string
        overallScore:
          type: integer
        analysis:
          $ref: '#/components/schemas/InterviewAnalysis'
        dimensions:
          type: array
          items:
            $ref: '#/components/schemas/ReviewDimension'
        questionAnalysis:
          type: array
          items:
            $ref: '#/components/schemas/QuestionAnalysis'
        improvementPlan:
          type: array
          items:
            $ref: '#/components/schemas/ImprovementPlan'
    ReviewListItem:
      type: object
      properties:
        id:
          type: string
        interviewId:
          type: string
        overallScore:
          type: integer
        date:
          type: string

    # ==================== 统计相关 ====================
...
    PageResponse:
      type: object
      properties:
        list:
          type: array
          items: {}
        total:
          type: integer
        page:
          type: integer
        size:
          type: integer
        pages:
          type: integer

    # ==================== 通用响应 ====================
    ApiResponse_Void:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        timestamp:
          type: integer
          format: int64
...
    ApiResponse_ResumeList:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/Resume'
        timestamp:
          type: integer
          format: int64
    ApiResponse_Resume:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/Resume'
        timestamp:
          type: integer
          format: int64
    ApiResponse_ResumeDetail:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/ResumeDetail'
        timestamp:
          type: integer
          format: int64
    ApiResponse_ResumeSuggestionList:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/ResumeSuggestion'
        timestamp:
          type: integer
          format: int64
    ApiResponse_OptimizeResume:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/OptimizeResumeResponse'
        timestamp:
          type: integer
          format: int64
    ApiResponse_StartSession:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/StartSessionResponse'
        timestamp:
          type: integer
          format: int64
    ApiResponse_SubmitAnswer:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/SubmitAnswerResponse'
        timestamp:
          type: integer
          format: int64
    ApiResponse_Hint:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/HintResponse'
        timestamp:
          type: integer
          format: int64
    ApiResponse_FinishSession:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/FinishSessionResponse'
        timestamp:
          type: integer
          format: int64
    ApiResponse_InterviewHistoryPage:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          allOf:
            - $ref: '#/components/schemas/PageResponse'
            - type: object
              properties:
                list:
                  type: array
                  items:
                    $ref: '#/components/schemas/Interview'
        timestamp:
          type: integer
          format: int64
    ApiResponse_InterviewDetail:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/InterviewDetail'
        timestamp:
          type: integer
          format: int64
    ApiResponse_InterviewQuestions:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/InterviewQuestions'
        timestamp:
          type: integer
          format: int64
    ApiResponse_ReviewListPage:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          allOf:
            - $ref: '#/components/schemas/PageResponse'
            - type: object
              properties:
                list:
                  type: array
                  items:
                    $ref: '#/components/schemas/ReviewListItem'
        timestamp:
          type: integer
          format: int64
    ApiResponse_InterviewReview:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/InterviewReview'
        timestamp:
          type: integer
          format: int64
...